requirements.txt
ninja>=1.10.2.4
nevergrad>=0.5.0
pytest>=7.1.3
pytest-html>=3.1.1
pandas>=1.5.0
pymongo>=4.2.0
graphviz>=0.20.1
seaborn>=0.12.0
tensorboard>=2.10.1
coloredlogs>=15.0.1
wandb>=0.13.5
gymnasium>=0.27.0
termcolor>=2.2.0
opencv-python>=4.7.0.68
pygame>=2.1.2
moviepy>=1.0.3
pyyaml>=6.0
tensorboard>=2.12.0
ray[default]>=2.3.0
codon-jit>=0.1.3